Add optional CGO pxlib backends
