+++ docs/WINDOWS_BUILD.md
@@ -128,16 +128,22 @@ The version information will appear in the Windows file properties dialog.
 
 ## Using in patris-export
 
-Once you have the pxlib DLL, you can build patris-export for Windows:
+Once you have pxlib, select one of the three supported native backends. The
+runtime-dynamic backend remains the default; direct CGO shared and static
+linking are opt-in build variants:
 
 ### On Windows:
-```bash
-go build -o patris-export.exe ./cmd/patris-export
+```powershell
+.\scripts\windows\Build-LocalWindows.ps1 -PxlibBackend dynamic
+.\scripts\windows\Build-LocalWindows.ps1 -PxlibBackend cgo -SkipPromote
+.\scripts\windows\Build-LocalWindows.ps1 -PxlibBackend cgo-static -SkipPromote
 ```
 
 ### Cross-compile from Linux:
 ```bash
-make build-windows
+./build.sh --target windows-cross --pxlib-backend dynamic
+./build.sh --target windows-cross --pxlib-backend cgo
+./build.sh --target windows-cross --pxlib-backend cgo-static
 ```
 
 Keep `libpxlib.dll` in the same directory as the patris-export executable for
@@ -147,3 +153,10 @@ missing. A foreground `.db` read reports a Patris Export native-runtime error
 and, by default, shows a Windows TaskDialog-style message with the checked
 paths. Set `PATRIS_EXPORT_NO_TASKDIALOG=1` for scripts or services that must
 never show a modal dialog.
+
+The `cgo` build imports pxlib normally at process start. The `cgo-static` build
+uses `libpxlib_static.a`, generated by the source-build helpers from the same
+pinned upstream pxlib commit, and does not package `libpxlib.dll`. MinGW runtime
+DLLs may still be required because the application itself uses CGO. For the
+complete backend contract and verification commands, see
+[Native pxlib backends](NATIVE-PXLIB-BACKENDS.md).
