+++ pkg/paradox/native_cgo_shared_test.go
@@ -1,0 +1,11 @@
+//go:build cgo && pxlib_cgo && !pxlib_cgo_static
+
+package paradox
+
+import "testing"
+
+func TestCGOSharedBackendIdentity(t *testing.T) {
+	if got := NativeBackend(); got != "cgo-shared" {
+		t.Fatalf("NativeBackend() = %q, want cgo-shared", got)
+	}
+}
