+++ scripts/build-pxlib-linux.sh
@@ -32,6 +32,10 @@ if [ ! -f "$PREFIX/include/paradox.h" ]; then
     cp "$BUILD_DIR"/include/*.h "$PREFIX/include/" 2>/dev/null || true
 fi
 find "$BUILD_DIR" -maxdepth 3 \( -name 'libpx*.a' -o -name 'libpx*.so*' \) -exec cp {} "$PREFIX/lib/" \; 2>/dev/null || true
+static_objects="$(find "$BUILD_DIR" -path '*/CMakeFiles/pxlib.dir/objects.a' -print -quit)"
+if [ -n "$static_objects" ]; then
+    cp "$static_objects" "$PREFIX/lib/libpx_static.a"
+fi
 
 if ls "$PREFIX/lib"/libpxlib.* >/dev/null 2>&1 && ! ls "$PREFIX/lib"/libpx.* >/dev/null 2>&1; then
     for lib in "$PREFIX/lib"/libpxlib.*; do
